refactor : BookDetails를 book/[isbn] 하위의 _components 폴더로 이동
