+++ src/app/book/[isbn]/_components/BookDetails/components/BookCover/index.tsx
