+++ src/app/book/[isbn]/_components/BookDetails/components/BookDescription/index.module.scss
