+++ src/app/book/[isbn]/_components/BookDetails/index.module.scss
