+++ src/app/book/[isbn]/_components/BookDetails/index.tsx
